# public/swagger.yaml
swagger: "2.0"
info:
    description: "This is a swagger documentation of API in HeartWood E-commerce"
    version: "1.0.7"
    title: "HeartWood E-commerce API Documentation"
    contact:
        email: "apiteam@swagger.io"
    license:
        name: "Apache 2.0"
        url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "localhost:3000"
basePath: "/api"
...
schemes:
    - "https"
    - "http"
paths:
...
    /cart/remove-item:
        post:
            tags:
                - "Cart"
            summary: "Remove item from user's cart"
            description: "Remove an item from user's cart"
            consumes:
                - "application/json"
            produces:
                - "application/json"
            parameters:
                - in: body
                  name: "userId"
                  description: "ID of the user"
                  required: true
                  schema:
                    type: integer
                - in: body
                  name: "productId"
                  description: "ID of the product"
                  required: true
                  schema:
                    type: integer
...